requirements.txt
fastapi 
uvicorn 
jinja2 
aiofiles 
Pillow 
tensorflow 
numpy
python-telegram-bot